Updated OgreOggSound which fixes the forced source code edit of last version. It was apparently a bug. http://www.ogre3d.org/addonforums/viewtopic.php?f=19&t=30189&p=102558#p102558 Added BOOST_DIR variable to toolchain

+++ contrib/ogre_ogg_sound/include/OgreOggSoundRecord.h
@@ -1,14 +1,14 @@
 /**
 * @file OgreOggSoundRecord.h
 * @author  Ian Stangoe
-* @version v1.23
+* @version v1.24
 *
 * @section LICENSE
 * 
 * This source file is part of OgreOggSound, an OpenAL wrapper library for   
 * use with the Ogre Rendering Engine.										 
 *                                                                           
-* Copyright (c) 2013 <Ian Stangoe>
+* Copyright (c) 2013 Ian Stangoe
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
@@ -60,13 +60,13 @@ namespace OgreOggSound
 	struct WAVEHEADER
 	{
 		char			szRIFF[4];
-		long			lRIFFSize;
+		int				lRIFFSize;
 		char			szWave[4];
 		char			szFmt[4];
-		long			lFmtSize;
+		int				lFmtSize;
 		wFormat			wfex;
 		char			szData[4];
-		long			lDataSize;
+		int				lDataSize;
 	};
 
 	//! Captures audio data
